requirements.txt
toml>=0.10.2
schedule>=1.1.0
pystray>=0.19.0
Pillow>=9.0.0
pywin32>=300; sys_platform == 'win32'
watchdog>=2.1.6